# swagger.yaml
basePath: /API
consumes:
    - application/json
definitions:
...
    ErrorResponse:
        description: 'Error response.'
        properties:
            responseCode:
                type: number
                format: double
                description: 'error code in system.'
            message:
                type: string
                description: 'Readble message.'
                x-nullable: true
        required:
            - responseCode
        type: object
...
    Login:
        description: 'Login schema model.'
        properties:
            email:
                type: string
            password:
                type: string
        required:
            - email
            - password
        type: object
    LocalServerInfo:
        description: 'Info about a local server.'
        properties:
            localServerId:
                type: string
                description: 'Local server id'
            displayName:
                type: string
                description: 'Display name'
        required:
            - localServerId
            - displayName
        type: object
    LoginLocalServer:
        description: "Extends login with local server selection,\r\ncase use owns more than one local server needs to know which server to connect."
        properties:
            email:
                type: string
            password:
                type: string
            localServerId:
                type: string
                description: "Local server to login to.\r\nif the user exists only in one local server ignore the field."
                x-nullable: true
        required:
            - email
            - password
        type: object
...
info:
    title: 'CASAnet - remote server API'
    version: 2.1.0
    description: 'Remote server for forwarding requests to locals home network. hosted in [GitHub repo](https://github.com/casanet/remote-server).'
    license:
        name: GNU
paths:
...
    /administration/auth/login:
        post:
            operationId: AdministrationLogin
            produces:
                - application/json
            responses:
                '201':
                    description: '2-fatore code sent'
                '204':
                    description: 'No content'
                '403':
                    description: 'Auth fail'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '422':
                    description: 'Invalid schema'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: 'Login to the administration system.'
            tags:
                - Administration
            security: []
            parameters:
                -
                    in: body
                    name: login
                    required: true
                    schema:
                        $ref: '#/definitions/Login'
    /administration/auth/login/tfa:
        post:
            operationId: AdministrationLoginTfa
            produces:
                - application/json
            responses:
                '204':
                    description: 'No content'
                '403':
                    description: 'Auth fail'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '422':
                    description: 'Invalid schema'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: '2-step verification login to the administration system.'
            tags:
                - Administration
            security: []
            parameters:
                -
                    in: body
                    name: login
                    required: true
                    schema:
                        $ref: '#/definitions/Login'
    /administration/auth/logout:
        post:
            operationId: AdministrationLogout
            produces:
                - application/json
            responses:
                '204':
                    description: 'No content'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: 'Logout manually from the administration system.'
            tags:
                - Administration
            security:
                -
                    adminAuth: []
            parameters: []
    /auth/login:
        post:
            operationId: Login
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: object
                '201':
                    description: '2-fatore code sent'
                '210':
                    description: 'select local server to connect to'
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/LocalServerInfo'
                '403':
                    description: 'Auth fail'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '422':
                    description: 'Invalid schema'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: "Login to local server via remote server channel.\r\nIf users exists in more then one local server, it returns status code 210 and the available user servers to select."
            tags:
                - Authentication
            security: []
            parameters:
                -
                    in: body
                    name: login
                    required: true
                    schema:
                        $ref: '#/definitions/LoginLocalServer'
    /auth/login/tfa:
        post:
            operationId: LoginTfa
            produces:
                - application/json
            responses:
                '204':
                    description: 'No content'
                '403':
                    description: 'Auth fail'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '422':
                    description: 'Invalid schema'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: '2-step verification login to local server via remote server channel.'
            tags:
                - Authentication
            security: []
            parameters:
                -
                    in: body
                    name: login
                    required: true
                    schema:
                        $ref: '#/definitions/LoginLocalServer'
    /auth/logout:
        post:
            operationId: Logout
            produces:
                - application/json
            responses:
                '204':
                    description: 'No content'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: 'Logout manually from remote and local server systems.'
            tags:
                - Authentication
            security:
                -
                    forwardAuth: []
            parameters: []
...
produces:
    - application/json
swagger: '2.0'
securityDefinitions:
    adminAuth:
        type: apiKey
        name: Cookie
        in: header
    forwardAuth:
        type: apiKey
        name: Cookie
        in: header
    iftttAuth:
        type: apiKey
        name: Cookie
        in: header
host: 'localhost:3001'
...
schemes:
    - http
    - https